Disable bundling for now, and remove cyclic imports
