+++ packages/next/src/constants.ts
@@ -1,0 +1,14 @@
+export const KIB = 1024;
+
+export const MIB = 1024 * KIB;
+
+/**
+ * The maximum size of a *compressed* edge function.
+ */
+export const EDGE_FUNCTION_SIZE_LIMIT = 4 * MIB;
+
+export const MAX_UNCOMPRESSED_LAMBDA_SIZE = 250 * MIB;
+
+export const LAMBDA_RESERVED_UNCOMPRESSED_SIZE = 2.5 * MIB;
+
+export const LAMBDA_RESERVED_COMPRESSED_SIZE = 250 * KIB;
