--- packages/next/src/edge-function-source/constants.ts
@@ -1,6 +1,0 @@
-import { MIB } from '../utils';
-
-/**
- * The maximum size of a *compressed* edge function.
- */
-export const EDGE_FUNCTION_SIZE_LIMIT = 4 * MIB;
