+++ packages/next/src/pretty-bytes.ts
@@ -1,0 +1,3 @@
+import bytes from 'bytes';
+
+export const prettyBytes = (n: number) => bytes(n, { unitSeparator: ' ' });
